# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/seasoning": {
            "get": {
                "tags": [
                    "Seasoning"
                ],
                "summary": "get seasonings",
                "operationId": "getSeasonings",
                "responses": {
                    "200": {
                        "description": "success,return array of seasonings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "product_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/seasonings": {
            "post": {
                "tags": [
                    "Seasoning"
                ],
                "summary": "store a seasoning",
                "operationId": "f3d53ce19dcaea6d622625a0993f0921",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "product name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "select category for product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success,return object of product",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "product_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/seasonings/{seasoning_id}": {
            "get": {
                "tags": [
                    "Seasoning"
                ],
                "summary": "get a seasoning by id",
                "operationId": "f37022213ff3aea49eeca8be5a56547a",
                "responses": {
                    "200": {
                        "description": "success,return object of seasoning",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "product_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            },
            "put": {
                "tags": [
                    "Seasoning"
                ],
                "summary": "update a seasoning",
                "operationId": "8a633ce21bd8b73a28454013473d9706",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "role name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "product_id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation error"
                    }
                }
            }
        },
        "/seasoning/{seasoning_id}": {
            "delete": {
                "tags": [
                    "Seasoning"
                ],
                "summary": "delete a seasoning",
                "operationId": "0c39653360a7b500bed463c25349242a",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation error"
                    }
                }
            }
        },
...
        "/userRoles": {
            "get": {
                "tags": [
                    "UserRole"
                ],
                "summary": "get all user's roles",
                "operationId": "9e04def149814393febd1a5d0ffd80c9",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success,return array of role objects",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/userAttachRole": {
            "post": {
                "tags": [
                    "UserRole"
                ],
                "summary": "attach a role to user",
                "operationId": "965490f03afdc64121ba35c574f58625",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        },
        "/userDetachRole": {
            "post": {
                "tags": [
                    "UserRole"
                ],
                "summary": "detach a role of user",
                "operationId": "34950eff818c29803ae591f72b93ebfe",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "role_id",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation errors"
                    }
                }
            }
        }
    }
}